ggml-dsp: split ggml-dsp.c into multiple files and cleanup
