+++ scripts/ggml-hexagon.cfg
@@ -23,9 +23,9 @@
 #
 [general]
 #version of ggml-hexagon.cpp on ARM-AP side
-version = "1.04"
+version = "1.05"
 #version of ggml-dsp.c on cDSP side
-ggmldsp_version = "0.61"
+ggmldsp_version = "0.62"
 
 #0: HEXAGON_BACKEND_QNNCPU
 #1: HEXAGON_BACKEND_QNNGPU
